+++ src/compiler/binder/mod.rs
@@ -6,5 +6,7 @@ mod lines_1000_1500;
 pub use lines_1000_1500::*;
 mod lines_1500_2000;
 pub use lines_1500_2000::*;
-mod lines_2000_3000;
-pub use lines_2000_3000::*;
+mod lines_2000_2500;
+pub use lines_2000_2500::*;
+mod lines_2500_3000;
+pub use lines_2500_3000::*;
