+++ src/lib.rs
@@ -285,22 +285,22 @@ pub use compiler::utilities::{
     get_language_variant, get_leftmost_access_expression, get_leftmost_expression,
     get_literal_text, get_locale_specific_message, get_object_flags, get_operator_associativity,
     get_operator_precedence, get_right_most_assigned_expression, get_root_declaration,
-    get_source_file_of_node, get_source_text_of_node_from_source_file, get_strict_option_value,
-    get_symbol_name_for_private_identifier, get_syntactic_modifier_flags,
-    get_text_of_identifier_or_literal, get_text_of_node_from_source_text,
-    get_text_of_property_name, has_dynamic_name, has_effective_readonly_modifier,
-    has_static_modifier, has_syntactic_modifier, is_access_expression, is_ambient_module,
-    is_any_import_or_re_export, is_assignment_expression, is_assignment_operator,
-    is_assignment_target, is_bindable_static_access_expression,
+    get_source_file_of_node, get_source_text_of_node_from_source_file,
+    get_span_of_token_at_position, get_strict_option_value, get_symbol_name_for_private_identifier,
+    get_syntactic_modifier_flags, get_text_of_identifier_or_literal,
+    get_text_of_node_from_source_text, get_text_of_property_name, get_token_pos_of_node,
+    has_dynamic_name, has_effective_readonly_modifier, has_static_modifier, has_syntactic_modifier,
+    is_access_expression, is_ambient_module, is_any_import_or_re_export, is_assignment_expression,
+    is_assignment_operator, is_assignment_target, is_bindable_static_access_expression,
     is_bindable_static_element_access_expression, is_block_or_catch_scoped,
     is_destructuring_assignment, is_dotted_name, is_dynamic_name, is_effective_module_declaration,
     is_entity_name_expression, is_enum_const, is_exports_identifier,
     is_expression_with_type_arguments_in_class_extends_clause, is_external_or_common_js_module,
     is_function_block, is_function_expression_or_arrow_function, is_global_scope_augmentation,
-    is_import_call, is_in_js_file, is_in_jsdoc, is_jsdoc_construct_signature, is_jsdoc_type_alias,
-    is_json_source_file, is_keyword, is_literal_import_type_node,
-    is_logical_or_coalescing_assignment_operator, is_module_augmentation_external,
-    is_module_exports_access_expression, is_object_literal_method,
+    is_identifier_name, is_import_call, is_in_js_file, is_in_jsdoc, is_in_top_level_context,
+    is_jsdoc_construct_signature, is_jsdoc_type_alias, is_json_source_file, is_keyword,
+    is_literal_import_type_node, is_logical_or_coalescing_assignment_operator,
+    is_module_augmentation_external, is_module_exports_access_expression, is_object_literal_method,
     is_object_literal_or_class_expression_method_or_accessor, is_prologue_directive,
     is_property_access_entity_name_expression, is_property_name_literal, is_prototype_access,
     is_push_or_unshift_identifier, is_require_call, is_signed_numeric_literal, is_static,
